[tailsampling] Efficient trace sampling based on attributes
The tail sampling processor in its original form has the ability to sample based on some attribute value, however, it lacks the ability to easily discard traces based on an attribute value. To do that we resorted to confusing and inefficient processing. e.g. When filtering out health endpoints the sampler would iterate over ALL the trace attributes for each policy defined policy (we currently have 5), not match anything, then fall through to the composite rate sampling policy to get our health endpoint capture down to about 1%.

This change adds a new policy evaluator to the tail sampling project to more directly and efficiently. It is called a logical OR attribute evaluator and its made up of one or more string or numeric attribute evaluators. If any of them match the rest of them are short-circuited. Very similar to the existing string and numeric filters but implemented differently. The trace attributes are iterated over no more than once (for this policy). A new final decision state was added called NotSampledFinal to support returning a NotSampled final decision. This wasn't possible before and this is how we are able to filter out health endpoints quickly. Metrics for the tail sampling processor were adjusted a bit to make more sense.